BACKPORT: UPSTREAM: sg_start_req(): make sure that there's not too many elements in iovec
unfortunately, allowing an arbitrary 16bit value means a possibility of
overflow in the calculation of total number of pages in bio_map_user_iov() -
we rely on there being no more than PAGE_SIZE members of sum in the
first loop there.  If that sum wraps around, we end up allocating
too small array of pointers to pages and it's easy to overflow it in
the second loop.

Bug: 35841297
X-Coverup: TINC (and there's no lumber cartel either)
Cc: stable@vger.kernel.org # way, way back
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
[toddpoynor@google.com: fixups for 3.10]
Change-Id: Iabb082942146ed9dc1087462bdb8e1e957b8f309
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>