ipv4: keep skb->dst around in presence of IP options
Andrey Konovalov got crashes in __ip_options_echo() when a NULL skb->dst
is accessed.

ipv4_pktinfo_prepare() should not drop the dst if (evil) IP options
are present.

We could refine the test to the presence of ts_needtime or srr,
but IP options are not often used, so let's be conservative.

Thanks to syzkaller team for finding this bug.

Upstream commit: 34b2cef20f19c87999fff3da4071e66937db9644
Fixes: d826eb14ecef ("ipv4: PKTINFO doesnt need dst reference")
Bug: 35805460
Signed-off-by: Eric Dumazet <edumazet@google.com>
Reported-by: Andrey Konovalov <andreyknvl@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Change-Id: I4428212665ba44b6f4a3d5999bcd8409a79bea04
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>